// api/resources/assets/client/src/app/models/CampaignList.ts
export class CampaignList {
  id: number;
  unique_script_id: string;
  created_by: number;
  campaign_name: string;
  domain_name: string;
  styles: string;
  delay: number;
  delay_before_start: number;
  loop: any;
  exit_pop_up: any;
  branding?: number;
  branding_id?: number;
  created_at: any;
  updated_at: any;
  deleted_at: any;
  is_active: string;
}