// api/resources/assets/client/src/app/models/exitPopUpsList.ts
export class ExitPopUpsList {
  id: number;
  created_by: number;
  name: string;
  header_text: string;
  header_background_color: string;
  header_text_color: string;
  semi_header_text: string;
  semi_header_text_color: string;
  body_background_color: string;
  deleted_at: string;
  created_at: string;
  updated_at: string;
  sticky_reviews?: any;
  cta_link_url: string;
  campaign_id: number;
  btn_size?: string;
  btn_text?: string;
  btn_color?: string;
  btn_text_color?: string;
}